requirements.txt
streamlit
torch
torchvision
Pillow
numpy
opencv-python-headless
streamlit-webrtc
av
